Refactoring entities.

+++ src/main/java/com/trl/library/controller/dto/AddressDTO.java
@@ -17,6 +17,6 @@ public class AddressDTO {
     private String city;
     private String street;
     private String houseNumber;
-    private Long postcode;
+    private Integer postcode;
 
 }
